feat(android): align DTO/domain/Room layer with NestJS API
§7.4 / §14.4.3. Synchronises the courier client with the new backend
contract — see docs/{auth,orders,couriers,photos,statistics}.md.

DTOs lose the `{success, message, data}` envelope: every response is a
plain object straight from Nest, with camelCase keys (`accessToken`,
`fullName`, `assignedAt`, …). All ids switch from `Long` to `String`
(UUIDs), `LoginResponse.user` becomes a `CourierProfileDto`
(`isActive`/`isPaused` available for §7.5), `OrderDto` carries the full
timestamp set (`createdAt`, nullable `assignedAt`/`pickedUpAt`/
`nearCustomerAt`/`deliveredAt`/`returnedAt`) plus an embedded
`photos: PhotoMetaDto[]`, and `StatisticsDto` matches
`CourierSelfStatsResponse` (period + named-period vocabulary).
`OrderStatus.NEW` joins the enum for completeness; courier transitions
still start at `ASSIGNED`. `Order.price` is intentionally absent
(courier responses strip it per §15.1).

`LoginResponse` no longer carries `expiresIn` — the new `JwtUtils`
helper reads the `exp` claim out of the access token and falls back to
15 minutes if decode fails. `logout` now sends `{ refreshToken }` and
swallows network errors so local tokens are always cleared. `refresh`
matches the camelCase contract.

Room schema bumped to v3 (destructive migration was already on) — id
columns become text, timestamp set widened, `UserEntity` gains
`isActive`/`isPaused`. `OrderDao` re-sorts by `createdAt DESC` for
active, `returnedAt DESC` for history. Repository implementations
adapt to the new envelope and forward the new query params (`from`/
`to`, `period`).

Courier registration is gone — accounts are created by admins
(see docs/couriers.md). `RegisterFragment`/`ViewModel`/`UiState` +
`fragment_register.xml` + `tvRegisterLink` + `action_login_to_register`
removed. The orders list FAB "Получить заказ" hidden — auto-assign on
the backend makes manual creation moot.

`PhotoFileManager` now keys files on `orderId: String` (UUID sanitised
for filesystem use). `OrderDetailsUiState` swaps `photoUrl` for
`lastUploadedPhotoId` since byte access is gated by auth and goes
through a separate endpoint.

+++ android/app/src/main/java/com/example/curier_mobile/presentation/photo/PhotoCaptureFragment.kt
@@ -134,7 +134,7 @@ class PhotoCaptureFragment : BaseFragment<FragmentPhotoCaptureBinding>() {
         // Create output file
         val photoFile = PhotoFileManager.createPhotoFile(
             requireContext(),
-            args.orderId.toLong()
+            args.orderId
         )
 
         val outputOptions = ImageCapture.OutputFileOptions.Builder(photoFile).build()
